Update README.md
Harmonious PLP is a sleek and responsive Product Listing Page built using Next.js and modular CSS. It features a modern UI with hover effects, a toggleable favorite button, loading skeletons for improved UX, and a mobile-first responsive design that looks great on all devices. The project is designed to handle filtering and sorting functionality, making it ideal for e-commerce platforms or product showcases. Fully deployed on Netlify, it’s a perfect example of front-end skill, UI consistency, and performance optimization.